requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
chromadb==0.5.3
openai==1.40.0
redis==5.0.6
pydantic==2.8.2
apscheduler==3.10.4
httpx==0.27.0
python-dotenv==1.0.1
pytest==8.3.2
pytest-asyncio==0.24.0
langgraph==0.2.25
langgraph-checkpoint==1.0.12
langchain-core==0.3.0
langchain==0.3.0
python-jose[cryptography]==3.3.0
psycopg[binary]
psycopg-pool==3.2.1
neo4j==5.23.1
langfuse==2.36.0
pyyaml==6.0.2
